+++ pages/routes/test-route.tsx
@@ -1,0 +1,11 @@
+import Layout from '../../components/layout';
+
+const TestRoute = () => {
+  return (
+    <Layout>
+      <h2 className="invisible">Test Route</h2>
+    </Layout>
+  );
+};
+
+export default TestRoute;
